Overwrite support (#15)
* Overwrite support

* Version bump

+++ src/cli/index.ts
@@ -22,11 +22,17 @@ const argv = yargs(process.argv.slice(2))
       describe: "filename to save alongside the css file",
       default: "styles.ts",
     },
+    overwrite: {
+      type: "boolean",
+      describe: "should the output file be overwritten if it exists",
+      default: false,
+    },
   })
   .parseSync();
 
 const cssFile = argv.css;
 const outputFileName = argv.output;
+const overwrite = argv.overwrite;
 
 findFiles(cssFile).then((files) => {
   files.forEach((file) => {
@@ -35,6 +41,12 @@ findFiles(cssFile).then((files) => {
     const output = generateOutput(config, path.basename(file));
     const folder = path.dirname(file);
     const outputPath = path.join(folder, outputFileName);
+    const exists = fs.existsSync(outputPath);
+
+    if (exists && !overwrite) {
+      console.log(`File ${outputPath} already exists, skipping`);
+      return;
+    }
 
     fs.writeFileSync(outputPath, output);
     console.log(
